Se realizan cambios sobre archivos para desacoplar funcionalidades
