+++ v1/General.php
@@ -18,6 +18,7 @@ namespace services\v1;
 use JsonException;
 use services\master\connection\Executor;
 use services\master\Responses;
+use services\set\Constant;
 use services\set\Regular;
 
 /**
@@ -31,6 +32,60 @@ use services\set\Regular;
  */
 class General
 {
+    /**
+     * Action process
+     *
+     * This method is useful for execute various actions required
+     *
+     * @param string $method    name method dynamic
+     * @param array  $arguments for search one or various results
+     *
+     * @return mixed
+     * @throws JsonException
+     */
+    public function __call(string $method, array $arguments):array
+    {
+        $response = new Responses();
+        $object = '';
+        $array = json_decode($arguments[0], true, 512, JSON_THROW_ON_ERROR);
+        if ($method === 'actionProcess') {
+            $object = $arguments[2];
+        }
+        if (!isset($array['token'])) {
+            return $response->unauthorized();
+        }
+        $token = $array['token'];
+        $value = $this->findToken($token);
+        return $this->executeProcess($value, $arguments[0], $arguments[1], $object);
+    }
+
+    /**
+     * Execute process
+     *
+     * This method return result execute query for save patient
+     *
+     * @param array  $arrayToken data text field
+     * @param string $json       data text field
+     * @param string $option     for execute process
+     * @param string $object     for required object
+     *
+     * @return string | int | mixed
+     */
+    private function executeProcess(
+        array $arrayToken,
+        string $json,
+        string $option,
+        string $object
+    ):array {
+        include_once sprintf("model/%s.php", $object);
+        $response = new Responses();
+        $objectClass = $this->selectClass($object);
+        if ($arrayToken) {
+            return $objectClass->{$option.'Validate'}($json);
+        }
+        return $response->unauthorized(Constant::INVALID_TOKEN);
+    }
+
     /**
      * Method
      *
@@ -63,8 +118,7 @@ class General
             $class = trim('services\v1\model\ ') . $object;
             return new $class;
         }
-            include_once 'Validator.php';
-        return new Validator();
+        return new self();
     }
 
     /**
